# src/Api.Server/swagger/swagger.v1.0.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MyApi",
    "description": "An example application with OpenAPI, Swashbuckle, and API versioning.",
    "contact": {
      "name": "TRAN Le Khanh",
      "email": "khanhcan148@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "paths": {
    "/api/Languages": {
      "get": {
        "tags": [
          "Languages"
        ],
        "operationId": "Languages.GetLanguages",
        "parameters": [
          {
            "name": "Ids",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "Query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json; v=1.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Select2ItemResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json; v=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Languages/{id}": {
      "get": {
        "tags": [
          "Languages"
        ],
        "operationId": "Languages.GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json; v=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/Select2ItemResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json; v=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Languages/PublishMessage": {
      "post": {
        "tags": [
          "Languages"
        ],
        "operationId": "Languages.PublishMessage",
        "parameters": [
          {
            "name": "orderId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "Select2ItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "disabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}